requirements.txt
Flask==3.1.1
pillow==10.4.0
qdrant-client==1.15.1
torch==2.7.0
transformers==4.55.2